backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 缓存和任务队列
redis==5.0.1
celery==5.3.4

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 机器学习
torch==2.1.0
transformers==4.35.2
peft==0.6.2
datasets==2.14.6
accelerate==0.24.1

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 数据处理
pandas==2.1.3
numpy==1.24.4
openpyxl==3.1.2

# 配置管理
pydantic==2.5.0
pydantic-settings==2.0.3

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 工具库
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1

# 系统监控
psutil==5.9.6

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0